analyst badge for watchlist
