+++ centometer-client/app/dashboard/watchlist/page.tsx
@@ -1,14 +1,15 @@
 import { WatchlistTable } from "@/components/watchlist/watchlist-table";
 import { StockResult, columns } from "@/components/watchlist/columns";
 import { createClient } from "@/utils/supabase/server";
+import { Suspense } from "react";
 
 async function getData(): Promise<StockResult[]> {
-  
+
   const supabase = createClient();
   const {
     data: { user },
   } = await supabase.auth.getUser();
-  const id = user?.id  
+  const id = user?.id
   const url = process.env.NEXT_PUBLIC_LAMBDA_SEARCH_URL + "watchlistService?id="
 
   const response = await fetch(`${url + id}`, {
@@ -20,7 +21,6 @@ async function getData(): Promise<StockResult[]> {
     return [];
   }
   const json = await response.json();
-  
   return json.watchlist
 
 }
@@ -31,7 +31,11 @@ export default async function Page() {
     <>
       <div className="my-2 md:my-6 w-full">
         <h1 className="text-4xl font-bold">Watchlist</h1>
-        <WatchlistTable columns={columns} data={data} />
+        <Suspense>
+          <WatchlistTable columns={columns} data={data} />
+        </Suspense>
+
+
       </div>
     </>
   );
